// src/models/IoK8sApiSchedulingV1PriorityClassList/types.ts
import { IoK8sApiSchedulingV1PriorityClassModelOptions as IoK8sApiSchedulingV1PriorityClass } from '../IoK8sApiSchedulingV1PriorityClass';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiSchedulingV1PriorityClassListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiSchedulingV1PriorityClass>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}